impl: swtich between & flip card in editor
This allows for the ability to switch between cards and flip cards whilst in edit mode. This makes the editing processes a lot more streamlined, as you can quickly edit a large deck of cards without having to constantly enter and exit edit mode.

This code is pretty ugly and might need some cleaning up.