impl: Add user ID field & associated getter
This adds a user ID field to the user structure on the backend, allowing
for users to be referenced by their ID, rather than exposing their
email.

Also creates a stub for the GetUserByID method, which will make fetching
user information from the database a lot quicker. It still needs to be
implemented.

issue #52