impl: Implement edit mode privacy.
If the user is not the owner of the deck, then when they try to enter
edit more they will be kicked back out to view mode. In order to
implement this I had to fundamentally change how caching user
information is done. Instead of logging the user in on each page load,
the user token is cached in localStorage.

issue #53