chore: replace panics for getUserByID-get public decks
replaced panics for getUserByID and getting public decks in QueryDeckReq.

for getUserByID, the panic has been replaced with an error 404 returned. (the token is good, but can't find the user information within.)

for finding public decks, the panic has been replaced with sending the user to  a page claiming that no public decks are available if it errors in finding them.